requirements.txt
dashscope==1.20.10
python-dotenv==1.0.1
python_pptx==0.6.23
